\name{finndat}
\alias{finndat}
\docType{data}
\title{Data from Finney (1947)}
\description{Binomial data from a bioassay.}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{z}}{a numeric vector of doses.}
    \item{\code{den}}{a numeric vector of  binomial denominators.}
    \item{\code{y}}{binomial responses. }
  }
}
\source{
Finney, D.J. (1947).  \emph{Probit Analysis: A Statistical Treatment of the Sigmoid 
Response Curve}. Cambridge University Press, London and New York.
}
\keyword{datasets}
