\name{likelihoodAsy-package}
\alias{likelihoodAsy-package}
\alias{likelihoodAsy}
\docType{package}
\title{
The R likelihoodAsy Package
}
\description{
Some functions for likelihood asymptotics, based on the expository paper

Pierce, D.A. and Bellio, R. (2016) Modern likelihood-frequentist inference. Submitted for publication.
}
\details{
For a detailed introduction see the vignette (\code{browseVignettes("likelihoodAsy")}), which includes an
R script that carries out the examples.


}
\author{
Author: Ruggero Bellio and Donald A Pierce 

Maintainer: Ruggero Bellio \email{ruggero.bellio@uniud.it}
}
\keyword{package}
