\name{likelihoodAsy-package}
\alias{likelihoodAsy-package}
\alias{likelihoodAsy}
\docType{package}
\title{
The R likelihoodAsy Package
}
\description{
Some functions for likelihood asymptotics, based on the expository paper

Pierce, D.A. and Bellio, R. (2015) Modern likelihood-frequentist inference. In preparation.
}
\details{
For a more detailed introduction see the vignette (\code{vignette("likelihoodAsy")}) or the documentation pages.


}
\author{
Author: Ruggero Bellio and Donald A Pierce 

Maintainer: Ruggero Bellio \email{ruggero.bellio@uniud.it}
}
\keyword{package}
